Remove baggage from textmap propagation; inject and extract baggage manually from headers inside of contrib
