+++ ddtrace/tracer/spancontext.go
@@ -124,13 +124,6 @@ type spanContextV1Adapter interface {
 	Tags() map[string]string
 }
 
-func (c *SpanContext) useForDistrTracing() bool {
-	if c == nil {
-		return false
-	}
-	return !c.traceID.Empty() && c.spanID > 0
-}
-
 // FromGenericCtx converts a ddtrace.SpanContext to a *SpanContext, which can be used
 // to start child spans.
 func FromGenericCtx(c ddtrace.SpanContext) *SpanContext {
@@ -169,12 +162,10 @@ func newSpanContext(span *Span, parent *SpanContext) *SpanContext {
 
 	context.traceID.SetLower(span.traceID)
 	if parent != nil {
-		if parent.useForDistrTracing() {
-			context.traceID.SetUpper(parent.traceID.Upper())
-			context.trace = parent.trace
-			context.origin = parent.origin
-			context.errors = parent.errors
-		}
+		context.traceID.SetUpper(parent.traceID.Upper())
+		context.trace = parent.trace
+		context.origin = parent.origin
+		context.errors = parent.errors
 		parent.ForeachBaggageItem(func(k, v string) bool {
 			context.setBaggageItem(k, v)
 			return true
