+++ ddtrace/tracer/textmap_test.go
@@ -2601,335 +2601,335 @@ func FuzzStringMutator(f *testing.F) {
 	})
 }
 
-func TestInjectBaggagePropagator(t *testing.T) {
+// func TestInjectBaggagePropagator(t *testing.T) {
 
-	assert := assert.New(t)
+// 	assert := assert.New(t)
 
-	propagator := NewPropagator(&PropagatorConfig{
-		BaggageHeader: "baggage",
-		TraceHeader:   "tid",
-		ParentHeader:  "pid",
-	})
-	tracer, err := newTracer(WithPropagator(propagator))
-	assert.NoError(err)
-	defer tracer.Stop()
+// 	propagator := NewPropagator(&PropagatorConfig{
+// 		BaggageHeader: "baggage",
+// 		TraceHeader:   "tid",
+// 		ParentHeader:  "pid",
+// 	})
+// 	tracer, err := newTracer(WithPropagator(propagator))
+// 	assert.NoError(err)
+// 	defer tracer.Stop()
 
-	root := tracer.StartSpan("web.request")
-	root.SetBaggageItem("foo", "bar")
-	ctx := root.Context()
-	headers := http.Header{}
+// 	root := tracer.StartSpan("web.request")
+// 	root.SetBaggageItem("foo", "bar")
+// 	ctx := root.Context()
+// 	headers := http.Header{}
 
-	carrier := HTTPHeadersCarrier(headers)
-	err = tracer.Inject(ctx, carrier)
-	assert.Nil(err)
+// 	carrier := HTTPHeadersCarrier(headers)
+// 	err = tracer.Inject(ctx, carrier)
+// 	assert.Nil(err)
 
-	assert.Equal(headers.Get("baggage"), "foo=bar")
-}
+// 	assert.Equal(headers.Get("baggage"), "foo=bar")
+// }
 
-func TestExtractBaggagePropagator(t *testing.T) {
-	tracer, err := newTracer()
-	assert.NoError(t, err)
-	defer tracer.Stop()
-	headers := TextMapCarrier{
-		DefaultTraceIDHeader:  "4",
-		DefaultParentIDHeader: "1",
-		DefaultBaggageHeader:  "foo=bar",
-	}
-	s, err := tracer.Extract(headers)
-	assert.NoError(t, err)
-	got := make(map[string]string)
-	s.ForeachBaggageItem(func(k, v string) bool {
-		got[k] = v
-		return true
-	})
-	assert.Len(t, got, 1)
-	assert.Equal(t, "bar", got["foo"])
-}
+// func TestExtractBaggagePropagator(t *testing.T) {
+// 	tracer, err := newTracer()
+// 	assert.NoError(t, err)
+// 	defer tracer.Stop()
+// 	headers := TextMapCarrier{
+// 		DefaultTraceIDHeader:  "4",
+// 		DefaultParentIDHeader: "1",
+// 		DefaultBaggageHeader:  "foo=bar",
+// 	}
+// 	s, err := tracer.Extract(headers)
+// 	assert.NoError(t, err)
+// 	got := make(map[string]string)
+// 	s.ForeachBaggageItem(func(k, v string) bool {
+// 		got[k] = v
+// 		return true
+// 	})
+// 	assert.Len(t, got, 1)
+// 	assert.Equal(t, "bar", got["foo"])
+// }
 
-func TestInjectBaggagePropagatorEncoding(t *testing.T) {
-	assert := assert.New(t)
+// func TestInjectBaggagePropagatorEncoding(t *testing.T) {
+// 	assert := assert.New(t)
 
-	propagator := NewPropagator(&PropagatorConfig{
-		BaggageHeader: "baggage",
-		TraceHeader:   "tid",
-		ParentHeader:  "pid",
-	})
-	tracer, err := newTracer(WithPropagator(propagator))
-	assert.NoError(err)
-	defer tracer.Stop()
+// 	propagator := NewPropagator(&PropagatorConfig{
+// 		BaggageHeader: "baggage",
+// 		TraceHeader:   "tid",
+// 		ParentHeader:  "pid",
+// 	})
+// 	tracer, err := newTracer(WithPropagator(propagator))
+// 	assert.NoError(err)
+// 	defer tracer.Stop()
 
-	root := tracer.StartSpan("web.request")
-	root.SetBaggageItem("userId", "Amélie")
-	root.SetBaggageItem("serverNode", "DF 28")
-	headers := http.Header{}
+// 	root := tracer.StartSpan("web.request")
+// 	root.SetBaggageItem("userId", "Amélie")
+// 	root.SetBaggageItem("serverNode", "DF 28")
+// 	headers := http.Header{}
 
-	carrier := HTTPHeadersCarrier(headers)
-	err = tracer.Inject(root.Context(), carrier)
-	assert.Nil(err)
-	actualBaggage := headers.Get("baggage")
-	// Instead of checking equality of the whole string, assert that both key/value pairs are present.
-	assert.Contains(actualBaggage, "userId=Am%C3%A9lie")
-	assert.Contains(actualBaggage, "serverNode=DF+28")
-}
+// 	carrier := HTTPHeadersCarrier(headers)
+// 	err = tracer.Inject(root.Context(), carrier)
+// 	assert.Nil(err)
+// 	actualBaggage := headers.Get("baggage")
+// 	// Instead of checking equality of the whole string, assert that both key/value pairs are present.
+// 	assert.Contains(actualBaggage, "userId=Am%C3%A9lie")
+// 	assert.Contains(actualBaggage, "serverNode=DF+28")
+// }
 
-func TestInjectBaggagePropagatorEncodingSpecialCharacters(t *testing.T) {
-	assert := assert.New(t)
+// func TestInjectBaggagePropagatorEncodingSpecialCharacters(t *testing.T) {
+// 	assert := assert.New(t)
 
-	propagator := NewPropagator(&PropagatorConfig{
-		BaggageHeader: "baggage",
-		TraceHeader:   "tid",
-		ParentHeader:  "pid",
-	})
-	tracer, err := newTracer(WithPropagator(propagator))
-	assert.NoError(err)
-	defer tracer.Stop()
+// 	propagator := NewPropagator(&PropagatorConfig{
+// 		BaggageHeader: "baggage",
+// 		TraceHeader:   "tid",
+// 		ParentHeader:  "pid",
+// 	})
+// 	tracer, err := newTracer(WithPropagator(propagator))
+// 	assert.NoError(err)
+// 	defer tracer.Stop()
 
-	root := tracer.StartSpan("web.request")
-	ctx := root.Context()
-	root.SetBaggageItem(",;\\()/:<=>?@[]{}", ",;\\")
-	headers := http.Header{}
+// 	root := tracer.StartSpan("web.request")
+// 	ctx := root.Context()
+// 	root.SetBaggageItem(",;\\()/:<=>?@[]{}", ",;\\")
+// 	headers := http.Header{}
 
-	carrier := HTTPHeadersCarrier(headers)
-	err = tracer.Inject(ctx, carrier)
-	assert.Nil(err)
+// 	carrier := HTTPHeadersCarrier(headers)
+// 	err = tracer.Inject(ctx, carrier)
+// 	assert.Nil(err)
 
-	assert.Equal(headers.Get("baggage"), "%2C%3B%5C%28%29%2F%3A%3C%3D%3E%3F%40%5B%5D%7B%7D=%2C%3B%5C")
-}
+// 	assert.Equal(headers.Get("baggage"), "%2C%3B%5C%28%29%2F%3A%3C%3D%3E%3F%40%5B%5D%7B%7D=%2C%3B%5C")
+// }
 
-func TestExtractBaggagePropagatorDecoding(t *testing.T) {
-	tracer, err := newTracer()
-	assert.NoError(t, err)
-	defer tracer.Stop()
-	headers := TextMapCarrier{
-		DefaultTraceIDHeader:  "4",
-		DefaultParentIDHeader: "1",
-		DefaultBaggageHeader:  "userId=Am%C3%A9lie,serverNode=DF+28",
-	}
-	s, err := tracer.Extract(headers)
-	assert.NoError(t, err)
-	got := make(map[string]string)
-	s.ForeachBaggageItem(func(k, v string) bool {
-		got[k] = v
-		return true
-	})
-	assert.Len(t, got, 2)
-	assert.Equal(t, "Amélie", got["userId"])
-	assert.Equal(t, "DF 28", got["serverNode"])
-}
+// func TestExtractBaggagePropagatorDecoding(t *testing.T) {
+// 	tracer, err := newTracer()
+// 	assert.NoError(t, err)
+// 	defer tracer.Stop()
+// 	headers := TextMapCarrier{
+// 		DefaultTraceIDHeader:  "4",
+// 		DefaultParentIDHeader: "1",
+// 		DefaultBaggageHeader:  "userId=Am%C3%A9lie,serverNode=DF+28",
+// 	}
+// 	s, err := tracer.Extract(headers)
+// 	assert.NoError(t, err)
+// 	got := make(map[string]string)
+// 	s.ForeachBaggageItem(func(k, v string) bool {
+// 		got[k] = v
+// 		return true
+// 	})
+// 	assert.Len(t, got, 2)
+// 	assert.Equal(t, "Amélie", got["userId"])
+// 	assert.Equal(t, "DF 28", got["serverNode"])
+// }
 
-func TestInjectBaggageMaxItems(t *testing.T) {
-	assert := assert.New(t)
+// func TestInjectBaggageMaxItems(t *testing.T) {
+// 	assert := assert.New(t)
 
-	propagator := NewPropagator(&PropagatorConfig{
-		BaggageHeader: "baggage",
-	})
-	tracer, err := newTracer(WithPropagator(propagator))
-	assert.NoError(err)
-	defer tracer.Stop()
+// 	propagator := NewPropagator(&PropagatorConfig{
+// 		BaggageHeader: "baggage",
+// 	})
+// 	tracer, err := newTracer(WithPropagator(propagator))
+// 	assert.NoError(err)
+// 	defer tracer.Stop()
 
-	root := tracer.StartSpan("web.request")
-	ctx := root.Context()
+// 	root := tracer.StartSpan("web.request")
+// 	ctx := root.Context()
 
-	for i := 0; i < baggageMaxItems+2; i++ {
-		iString := strconv.Itoa(i)
-		ctx.setBaggageItem("key"+iString, "val"+iString)
-	}
+// 	for i := 0; i < baggageMaxItems+2; i++ {
+// 		iString := strconv.Itoa(i)
+// 		ctx.setBaggageItem("key"+iString, "val"+iString)
+// 	}
 
-	headers := http.Header{}
+// 	headers := http.Header{}
 
-	carrier := HTTPHeadersCarrier(headers)
-	err = tracer.Inject(ctx, carrier)
-	assert.Nil(err)
+// 	carrier := HTTPHeadersCarrier(headers)
+// 	err = tracer.Inject(ctx, carrier)
+// 	assert.Nil(err)
 
-	headerValue := headers.Get("baggage")
-	items := strings.Split(headerValue, ",")
-	assert.Equal(baggageMaxItems, len(items))
-}
+// 	headerValue := headers.Get("baggage")
+// 	items := strings.Split(headerValue, ",")
+// 	assert.Equal(baggageMaxItems, len(items))
+// }
 
-func TestInjectBaggageMaxBytes(t *testing.T) {
-	assert := assert.New(t)
+// func TestInjectBaggageMaxBytes(t *testing.T) {
+// 	assert := assert.New(t)
 
-	propagator := NewPropagator(&PropagatorConfig{
-		BaggageHeader: "baggage",
-	})
-	tracer, err := newTracer(WithPropagator(propagator))
-	assert.NoError(err)
-	defer tracer.Stop()
+// 	propagator := NewPropagator(&PropagatorConfig{
+// 		BaggageHeader: "baggage",
+// 	})
+// 	tracer, err := newTracer(WithPropagator(propagator))
+// 	assert.NoError(err)
+// 	defer tracer.Stop()
 
-	root := tracer.StartSpan("web.request")
-	ctx := root.Context()
+// 	root := tracer.StartSpan("web.request")
+// 	ctx := root.Context()
 
-	baggageItems := map[string]string{
-		"key0": "o",
-		"key1": strings.Repeat("a", baggageMaxBytes/3),
-		"key2": strings.Repeat("b", baggageMaxBytes/3),
-		"key3": strings.Repeat("c", baggageMaxBytes/3),
-	}
+// 	baggageItems := map[string]string{
+// 		"key0": "o",
+// 		"key1": strings.Repeat("a", baggageMaxBytes/3),
+// 		"key2": strings.Repeat("b", baggageMaxBytes/3),
+// 		"key3": strings.Repeat("c", baggageMaxBytes/3),
+// 	}
 
-	ctx.baggage = baggageItems
-	headers := http.Header{}
+// 	ctx.baggage = baggageItems
+// 	headers := http.Header{}
 
-	carrier := HTTPHeadersCarrier(headers)
-	err = tracer.Inject(ctx, carrier)
-	assert.Nil(err)
+// 	carrier := HTTPHeadersCarrier(headers)
+// 	err = tracer.Inject(ctx, carrier)
+// 	assert.Nil(err)
 
-	headerValue := headers.Get("baggage")
-	headerSize := len([]byte(headerValue))
-	assert.LessOrEqual(headerSize, baggageMaxBytes)
-}
+// 	headerValue := headers.Get("baggage")
+// 	headerSize := len([]byte(headerValue))
+// 	assert.LessOrEqual(headerSize, baggageMaxBytes)
+// }
 
-func TestExtractBaggagePropagatorMalformedHeader(t *testing.T) {
-	t.Run("missing equal sign", func(t *testing.T) {
-		tracer, err := newTracer()
-		assert.NoError(t, err)
-		defer tracer.Stop()
-		headers := TextMapCarrier{
-			DefaultTraceIDHeader:  "4",
-			DefaultParentIDHeader: "1",
-			DefaultBaggageHeader:  "key1,key2=value2",
-		}
-		s, err := tracer.Extract(headers)
-		assert.NoError(t, err)
-		// since the header is malformed, we should not have any baggage items
-		got := make(map[string]string)
-		s.ForeachBaggageItem(func(k, v string) bool {
-			got[k] = v
-			return true
-		})
-		assert.Len(t, got, 0)
-	})
-	t.Run("missing value", func(t *testing.T) {
-		tracer, err := newTracer()
-		assert.NoError(t, err)
-		defer tracer.Stop()
-		headers := TextMapCarrier{
-			DefaultTraceIDHeader:  "4",
-			DefaultParentIDHeader: "1",
-			DefaultBaggageHeader:  "key1=value1,key2=",
-		}
-		s, err := tracer.Extract(headers)
-		assert.NoError(t, err)
-		// since the header is malformed, we should not have any baggage items
-		got := make(map[string]string)
-		s.ForeachBaggageItem(func(k, v string) bool {
-			got[k] = v
-			return true
-		})
-		assert.Len(t, got, 0)
-	})
-	t.Run("missing key", func(t *testing.T) {
-		tracer, err := newTracer()
-		assert.NoError(t, err)
-		defer tracer.Stop()
-		headers := TextMapCarrier{
-			DefaultTraceIDHeader:  "4",
-			DefaultParentIDHeader: "1",
-			DefaultBaggageHeader:  "key1=value1,=value2",
-		}
-		s, err := tracer.Extract(headers)
-		assert.NoError(t, err)
-		// since the header is malformed, we should not have any baggage items
-		got := make(map[string]string)
-		s.ForeachBaggageItem(func(k, v string) bool {
-			got[k] = v
-			return true
-		})
-		assert.Len(t, got, 0)
-	})
-	t.Run("missing key and value", func(t *testing.T) {
-		tracer, err := newTracer()
-		assert.NoError(t, err)
-		defer tracer.Stop()
-		headers := TextMapCarrier{
-			DefaultTraceIDHeader:  "4",
-			DefaultParentIDHeader: "1",
-			DefaultBaggageHeader:  "=,key1=value1",
-		}
-		s, err := tracer.Extract(headers)
-		assert.NoError(t, err)
-		// since the header is malformed, we should not have any baggage items
-		got := make(map[string]string)
-		s.ForeachBaggageItem(func(k, v string) bool {
-			got[k] = v
-			return true
-		})
-		assert.Len(t, got, 0)
-	})
-	t.Run("missing key-value pair", func(t *testing.T) {
-		tracer, err := newTracer()
-		assert.NoError(t, err)
-		defer tracer.Stop()
-		headers := TextMapCarrier{
-			DefaultTraceIDHeader:  "4",
-			DefaultParentIDHeader: "1",
-			DefaultBaggageHeader:  "key1=value1,",
-		}
-		s, err := tracer.Extract(headers)
-		assert.NoError(t, err)
-		// since the header is malformed, we should not have any baggage items
-		got := make(map[string]string)
-		s.ForeachBaggageItem(func(k, v string) bool {
-			got[k] = v
-			return true
-		})
-		assert.Len(t, got, 0)
-	})
-}
+// func TestExtractBaggagePropagatorMalformedHeader(t *testing.T) {
+// 	t.Run("missing equal sign", func(t *testing.T) {
+// 		tracer, err := newTracer()
+// 		assert.NoError(t, err)
+// 		defer tracer.Stop()
+// 		headers := TextMapCarrier{
+// 			DefaultTraceIDHeader:  "4",
+// 			DefaultParentIDHeader: "1",
+// 			DefaultBaggageHeader:  "key1,key2=value2",
+// 		}
+// 		s, err := tracer.Extract(headers)
+// 		assert.NoError(t, err)
+// 		// since the header is malformed, we should not have any baggage items
+// 		got := make(map[string]string)
+// 		s.ForeachBaggageItem(func(k, v string) bool {
+// 			got[k] = v
+// 			return true
+// 		})
+// 		assert.Len(t, got, 0)
+// 	})
+// 	t.Run("missing value", func(t *testing.T) {
+// 		tracer, err := newTracer()
+// 		assert.NoError(t, err)
+// 		defer tracer.Stop()
+// 		headers := TextMapCarrier{
+// 			DefaultTraceIDHeader:  "4",
+// 			DefaultParentIDHeader: "1",
+// 			DefaultBaggageHeader:  "key1=value1,key2=",
+// 		}
+// 		s, err := tracer.Extract(headers)
+// 		assert.NoError(t, err)
+// 		// since the header is malformed, we should not have any baggage items
+// 		got := make(map[string]string)
+// 		s.ForeachBaggageItem(func(k, v string) bool {
+// 			got[k] = v
+// 			return true
+// 		})
+// 		assert.Len(t, got, 0)
+// 	})
+// 	t.Run("missing key", func(t *testing.T) {
+// 		tracer, err := newTracer()
+// 		assert.NoError(t, err)
+// 		defer tracer.Stop()
+// 		headers := TextMapCarrier{
+// 			DefaultTraceIDHeader:  "4",
+// 			DefaultParentIDHeader: "1",
+// 			DefaultBaggageHeader:  "key1=value1,=value2",
+// 		}
+// 		s, err := tracer.Extract(headers)
+// 		assert.NoError(t, err)
+// 		// since the header is malformed, we should not have any baggage items
+// 		got := make(map[string]string)
+// 		s.ForeachBaggageItem(func(k, v string) bool {
+// 			got[k] = v
+// 			return true
+// 		})
+// 		assert.Len(t, got, 0)
+// 	})
+// 	t.Run("missing key and value", func(t *testing.T) {
+// 		tracer, err := newTracer()
+// 		assert.NoError(t, err)
+// 		defer tracer.Stop()
+// 		headers := TextMapCarrier{
+// 			DefaultTraceIDHeader:  "4",
+// 			DefaultParentIDHeader: "1",
+// 			DefaultBaggageHeader:  "=,key1=value1",
+// 		}
+// 		s, err := tracer.Extract(headers)
+// 		assert.NoError(t, err)
+// 		// since the header is malformed, we should not have any baggage items
+// 		got := make(map[string]string)
+// 		s.ForeachBaggageItem(func(k, v string) bool {
+// 			got[k] = v
+// 			return true
+// 		})
+// 		assert.Len(t, got, 0)
+// 	})
+// 	t.Run("missing key-value pair", func(t *testing.T) {
+// 		tracer, err := newTracer()
+// 		assert.NoError(t, err)
+// 		defer tracer.Stop()
+// 		headers := TextMapCarrier{
+// 			DefaultTraceIDHeader:  "4",
+// 			DefaultParentIDHeader: "1",
+// 			DefaultBaggageHeader:  "key1=value1,",
+// 		}
+// 		s, err := tracer.Extract(headers)
+// 		assert.NoError(t, err)
+// 		// since the header is malformed, we should not have any baggage items
+// 		got := make(map[string]string)
+// 		s.ForeachBaggageItem(func(k, v string) bool {
+// 			got[k] = v
+// 			return true
+// 		})
+// 		assert.Len(t, got, 0)
+// 	})
+// }
 
-func TestExtractOnlyBaggage(t *testing.T) {
-	t.Setenv("DD_TRACE_PROPAGATION_STYLE", "baggage")
-	headers := TextMapCarrier(map[string]string{
-		"baggage": "foo=bar,baz=qux",
-	})
+// func TestExtractOnlyBaggage(t *testing.T) {
+// 	t.Setenv("DD_TRACE_PROPAGATION_STYLE", "baggage")
+// 	headers := TextMapCarrier(map[string]string{
+// 		"baggage": "foo=bar,baz=qux",
+// 	})
 
-	tracer, err := newTracer()
-	assert.NoError(t, err)
-	defer tracer.Stop()
+// 	tracer, err := newTracer()
+// 	assert.NoError(t, err)
+// 	defer tracer.Stop()
 
-	ctx, err := tracer.Extract(headers)
-	assert.Nil(t, err)
+// 	ctx, err := tracer.Extract(headers)
+// 	assert.Nil(t, err)
 
-	// Expect tracer.Extract to return spanContext with 0 tID; it's the caller of tracer.Extract's responsibility to catch this (e.g, func startSpan)
-	assert.Equal(t, traceIDFrom64Bits(0), ctx.traceID)
+// 	// Expect tracer.Extract to return spanContext with 0 tID; it's the caller of tracer.Extract's responsibility to catch this (e.g, func startSpan)
+// 	assert.Equal(t, traceIDFrom64Bits(0), ctx.traceID)
 
-	got := make(map[string]string)
-	ctx.ForeachBaggageItem(func(k, v string) bool {
-		got[k] = v
-		return true
-	})
-	assert.Len(t, got, 2)
-	assert.Equal(t, "bar", got["foo"])
-	assert.Equal(t, "qux", got["baz"])
-}
+// 	got := make(map[string]string)
+// 	ctx.ForeachBaggageItem(func(k, v string) bool {
+// 		got[k] = v
+// 		return true
+// 	})
+// 	assert.Len(t, got, 2)
+// 	assert.Equal(t, "bar", got["foo"])
+// 	assert.Equal(t, "qux", got["baz"])
+// }
 
-func TestExtractBaggageFirstThenDatadog(t *testing.T) {
-	t.Setenv("DD_TRACE_PROPAGATION_STYLE", "baggage,datadog")
+// func TestExtractBaggageFirstThenDatadog(t *testing.T) {
+// 	t.Setenv("DD_TRACE_PROPAGATION_STYLE", "baggage,datadog")
 
-	// headers that include both baggage and valid Datadog trace headers.
-	headers := TextMapCarrier(map[string]string{
-		"baggage":             "item=xyz",
-		DefaultTraceIDHeader:  "12345",
-		DefaultParentIDHeader: "67890",
-		DefaultPriorityHeader: "1",
-	})
+// 	// headers that include both baggage and valid Datadog trace headers.
+// 	headers := TextMapCarrier(map[string]string{
+// 		"baggage":             "item=xyz",
+// 		DefaultTraceIDHeader:  "12345",
+// 		DefaultParentIDHeader: "67890",
+// 		DefaultPriorityHeader: "1",
+// 	})
 
-	tracer, err := newTracer()
-	assert.NoError(t, err)
-	defer tracer.Stop()
+// 	tracer, err := newTracer()
+// 	assert.NoError(t, err)
+// 	defer tracer.Stop()
 
-	ctx, err := tracer.Extract(headers)
-	assert.NoError(t, err)
+// 	ctx, err := tracer.Extract(headers)
+// 	assert.NoError(t, err)
 
-	// Even though baggage appears first, the trace ID must come from Datadog.
-	expectedTraceID := traceIDFrom64Bits(12345)
-	assert.Equal(t, expectedTraceID, ctx.traceID)
-	assert.Equal(t, uint64(67890), ctx.spanID)
+// 	// Even though baggage appears first, the trace ID must come from Datadog.
+// 	expectedTraceID := traceIDFrom64Bits(12345)
+// 	assert.Equal(t, expectedTraceID, ctx.traceID)
+// 	assert.Equal(t, uint64(67890), ctx.spanID)
 
-	got := make(map[string]string)
-	ctx.ForeachBaggageItem(func(k, v string) bool {
-		got[k] = v
-		return true
-	})
-	assert.Len(t, got, 1)
-	assert.Equal(t, "xyz", got["item"])
-}
+// 	got := make(map[string]string)
+// 	ctx.ForeachBaggageItem(func(k, v string) bool {
+// 		got[k] = v
+// 		return true
+// 	})
+// 	assert.Len(t, got, 1)
+// 	assert.Equal(t, "xyz", got["item"])
+// }
